refactor: move support_matrix.py from docs/ to utils/
Moves src/unilab/docs/support_matrix.py to src/unilab/utils/support_matrix.py
and updates all import sites. Eliminates the anti-pattern of keeping runtime
Python modules inside a docs/ package under src/unilab/.

Relates to agent/src-unilab-directory-evaluation.md issue #6 (P1).
