--- src/unilab/docs/__init__.py
@@ -1,1 +1,0 @@
-"""Documentation helpers derived from repository facts."""
